compiler: Close c functions with no body
For whatever reason, trying to insert `0` into a `vector<size_t>` causes a memory segmentation fault, which will occour in the specific situation that the first statement in the ir is a code block for a block with no statements inside. For now, I've added a special case to deal with this.